+++ components/src/lib.rs
@@ -3,6 +3,7 @@
 
 mod accordion;
 mod button;
+mod drag_drop;
 mod dropdown;
 mod external_link;
 mod gesture_area;
@@ -18,6 +19,7 @@ mod tooltip;
 
 pub use accordion::*;
 pub use button::*;
+pub use drag_drop::*;
 pub use dropdown::*;
 pub use external_link::*;
 pub use gesture_area::*;
